fix(wasm_loader): cast length value to uint64 for array initialization size calculation

+++ core/iwasm/interpreter/wasm_loader.c
@@ -1084,8 +1084,9 @@ load_init_expr(WASMModule *module, const uint8 **p_buf, const uint8 *buf_end,
                                     goto fail;
                                 }
 
-                                size = sizeof(WASMArrayNewInitValues)
-                                       + sizeof(WASMValue) * len_val.i32;
+                                size =
+                                    sizeof(WASMArrayNewInitValues)
+                                    + sizeof(WASMValue) * (uint64)len_val.i32;
                                 if (!(array_init_values = loader_malloc(
                                           size, error_buf, error_buf_size))) {
                                     goto fail;
